requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
google-genai==0.3.0
Werkzeug==3.0.1
Alembic==1.13.1
SQLAlchemy==2.0.23
gunicorn==22.0.0
groq=1.0.0
